+++ PEsample/src/api/questionaire/QuestionairePOSTService.java
@@ -21,38 +21,40 @@ import javax.ws.rs.core.Response;
 
 import com.vgu.sqm.questionnaire.Configuration;
 
+import model.Answers;
+
 @Path("/questionaire")
 public class QuestionairePOSTService {
-	@Path("/service")
+	@Path("/answers")
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Response submitQuestionaire(questionaireAnswers Answers)throws SQLException, NamingException{
+	public Response submitQuestionaire(Answers answers)throws SQLException, NamingException{
 		Connection db= Configuration.getAcademiaConnection();
 		try {
 			PreparedStatement st = db.prepareStatement("call StoreQuestionnaireDetail(\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\","
 					+ "\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\",\"?\")");
-			st.setString(1, Answers.getCname());
-			st.setString(2, Answers.getLname());
-			st.setString(3, Answers.getAttend());
-			st.setString(4, Answers.getGender());
-			st.setString(5, Answers.getQ1());
-			st.setString(6, Answers.getQ2());
-			st.setString(7, Answers.getQ3());
-			st.setString(8, Answers.getQ4());
-			st.setString(9, Answers.getQ5());
-			st.setString(10, Answers.getQ6());
-			st.setString(11, Answers.getQ7());
-			st.setString(12, Answers.getQ8());
-			st.setString(13, Answers.getQ9());
-			st.setString(14, Answers.getQ10());
-			st.setString(15, Answers.getQ11());
-			st.setString(16, Answers.getQ12());
-			st.setString(17, Answers.getQ13());
-			st.setString(18, Answers.getQ14());
-			st.setString(19, Answers.getQ15());
-			st.setString(20, Answers.getQ16());
-			st.setString(21, Answers.getQ17());
-			st.setString(22, Answers.getQ18());
+			st.setString(1, answers.getCname());
+			st.setString(2, answers.getLname());
+			st.setString(3, answers.getAttend());
+			st.setString(4, answers.getGender());
+			st.setString(5, answers.getQ1());
+			st.setString(6, answers.getQ2());
+			st.setString(7, answers.getQ3());
+			st.setString(8, answers.getQ4());
+			st.setString(9, answers.getQ5());
+			st.setString(10, answers.getQ6());
+			st.setString(11, answers.getQ7());
+			st.setString(12, answers.getQ8());
+			st.setString(13, answers.getQ9());
+			st.setString(14, answers.getQ10());
+			st.setString(15, answers.getQ11());
+			st.setString(16, answers.getQ12());
+			st.setString(17, answers.getQ13());
+			st.setString(18, answers.getQ14());
+			st.setString(19, answers.getQ15());
+			st.setString(20, answers.getQ16());
+			st.setString(21, answers.getQ17());
+			st.setString(22, answers.getQ18());
 			st.executeQuery();
 			return Response.ok().entity("New academic year successfully inserted").build();
 		}catch(SQLException e) {
