--- PEsample/src/api/questionaire/questionaireAnswers.java
@@ -1,162 +1,0 @@
-package api.questionaire;
-
-
-public class questionaireAnswers {
-	private String cname;
-	private String lname;
-	private String gender;
-	private String attend;
-	private String q1;
-	private String q2;
-	private String q3;
-	private String q4;
-	private String q5;
-	private String q6;
-	private String q7;
-	private String q8;
-	private String q9;
-	private String q10;
-	private String q11;
-	private String q12;
-	private String q13;
-	private String q14;
-	private String q15;
-	private String q16;
-	private String q17;
-	private String q18;
-	public String getCname() {
-		return cname;
-	}
-	public void setCname(String cname) {
-		this.cname = cname;
-	}
-	public String getLname() {
-		return lname;
-	}
-	public void setLname(String lname) {
-		this.lname = lname;
-	}
-	public String getGender() {
-		return gender;
-	}
-	public void setGender(String gender) {
-		this.gender = gender;
-	}
-	public String getAttend() {
-		return attend;
-	}
-	public void setAttend(String attend) {
-		this.attend = attend;
-	}
-	public String getQ1() {
-		return q1;
-	}
-	public void setQ1(String q1) {
-		this.q1 = q1;
-	}
-	public String getQ2() {
-		return q2;
-	}
-	public void setQ2(String q2) {
-		this.q2 = q2;
-	}
-	public String getQ3() {
-		return q3;
-	}
-	public void setQ3(String q3) {
-		this.q3 = q3;
-	}
-	public String getQ4() {
-		return q4;
-	}
-	public void setQ4(String q4) {
-		this.q4 = q4;
-	}
-	public String getQ5() {
-		return q5;
-	}
-	public void setQ5(String q5) {
-		this.q5 = q5;
-	}
-	public String getQ6() {
-		return q6;
-	}
-	public void setQ6(String q6) {
-		this.q6 = q6;
-	}
-	public String getQ7() {
-		return q7;
-	}
-	public void setQ7(String q7) {
-		this.q7 = q7;
-	}
-	public String getQ8() {
-		return q8;
-	}
-	public void setQ8(String q8) {
-		this.q8 = q8;
-	}
-	public String getQ9() {
-		return q9;
-	}
-	public void setQ9(String q9) {
-		this.q9 = q9;
-	}
-	public String getQ10() {
-		return q10;
-	}
-	public void setQ10(String q10) {
-		this.q10 = q10;
-	}
-	public String getQ11() {
-		return q11;
-	}
-	public void setQ11(String q11) {
-		this.q11 = q11;
-	}
-	public String getQ12() {
-		return q12;
-	}
-	public void setQ12(String q12) {
-		this.q12 = q12;
-	}
-	public String getQ13() {
-		return q13;
-	}
-	public void setQ13(String q13) {
-		this.q13 = q13;
-	}
-	public String getQ14() {
-		return q14;
-	}
-	public void setQ14(String q14) {
-		this.q14 = q14;
-	}
-	public String getQ15() {
-		return q15;
-	}
-	public void setQ15(String q15) {
-		this.q15 = q15;
-	}
-	public String getQ16() {
-		return q16;
-	}
-	public void setQ16(String q16) {
-		this.q16 = q16;
-	}
-	public String getQ17() {
-		return q17;
-	}
-	public void setQ17(String q17) {
-		this.q17 = q17;
-	}
-	public String getQ18() {
-		return q18;
-	}
-	public void setQ18(String q18) {
-		this.q18 = q18;
-	}
-	
-	
-}
-
